requirements.txt
fastapi
SQLAlchemy
PyMySQL
colorama
python-jose
uvicorn
python-multipart
